backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
requests
boto3
django-storages
razorpay
gunicorn
python-dotenv
social-auth-app-django
